pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "checkoutdesignator"
version = "0.1.0"
description = "CheckoutDesignator – modern SK Games checkout and preorder system"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "SK R&D" }
]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "sqlmodel>=0.0.21",
  "sqlalchemy>=2.0",
  "python-multipart>=0.0.6",
  "typer>=0.12.3",
  "rich>=13.7.0",
  "tabulate>=0.9.0",
  "pydantic-settings>=2.3",
  "httpx>=0.26",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "anyio>=4.2",
  "pytest-cov>=4.1"
]

[project.scripts]
checkout-designator = "checkoutdesignator.cli:app"

[tool.setuptools]
packages = ["checkoutdesignator"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
